+++ iotw-backend/src/main/java/org/example/utils/Const.java
@@ -16,10 +16,10 @@ public class Const {
     public static final int ORDER_CORS = -102;
     public static final int ORDER_LIMIT = -101;
 
-    public static final int RULE_USER = 1;
+    public static final Short RULE_USER = 1;
 
-    public static final int IS_NOT_DELETED = 0;
-    public static final int IS_DELETED = 1;
-    public static final int SEX_MAN = 1;
-    public static final int SEX_WOMAN = 0;
+    public static final Short IS_NOT_DELETED = 0;
+    public static final Short IS_DELETED = 1;
+    public static final Short SEX_MAN = 1;
+    public static final Short SEX_WOMAN = 0;
 }
